third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-all.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
#initial { all: initial; }
#inherit { all: inherit; }
#invalid { all: green; }
#notOnlyInitial { all: initial; color: green; }
#notOnlyInherit { all: inherit; color: green; }
</style>
<div id="initial"></div>
<div id="inherit"></div>
<div id="invalid"></div>
<div id="notOnlyInitial"></div>
<div id="notOnlyInherit"></div>
<script>
description('Test calling getPropertyValue on computed styles for all shorthand property.');

shouldBeEqualToString('getComputedStyle(document.getElementById("initial")).all', "");
shouldBeEqualToString('getComputedStyle(document.getElementById("inherit")).all', "");
shouldBeEqualToString('getComputedStyle(document.getElementById("invalid")).all', "");
shouldBeEqualToString('getComputedStyle(document.getElementById("notOnlyInitial")).all', "");
shouldBeEqualToString('getComputedStyle(document.getElementById("notOnlyInherit")).all', "");
</script>
