third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-all-expected.txt

Test calling getPropertyValue on computed styles for all shorthand property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL getComputedStyle(document.getElementById("initial")).all should be  (of type string). Was undefined (of type undefined).
FAIL getComputedStyle(document.getElementById("inherit")).all should be  (of type string). Was undefined (of type undefined).
FAIL getComputedStyle(document.getElementById("invalid")).all should be  (of type string). Was undefined (of type undefined).
FAIL getComputedStyle(document.getElementById("notOnlyInitial")).all should be  (of type string). Was undefined (of type undefined).
FAIL getComputedStyle(document.getElementById("notOnlyInherit")).all should be  (of type string). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
